api/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-jose
requests
reportlab
passlib[bcrypt]
